requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiofiles==23.2.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Validation
pydantic==2.7.4
email-validator==2.1.1

# Auth & Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8

# Environment
python-dotenv==1.0.1